# Implement Bernoulli distribution