--- a/src/distributions/bernoulli.rs
+++ b/src/distributions/bernoulli.rs
@@ -19,7 +19,7 @@ use distributions::Distribution;
 #[derive(Clone, Copy, Debug)]
 pub struct Bernoulli {
     /// Probability of success, relative to the maximal integer.
-    p_int: u32,
+    p: f64,
 }
 
 impl Bernoulli {
@@ -32,16 +32,14 @@ impl Bernoulli {
     pub fn new(p: f64) -> Bernoulli {
         assert!(p >= 0.0, "Bernoulli::new called with p < 0");
         assert!(p <= 1.0, "Bernoulli::new called with p > 1");
-        let p_int = (p * f64::from(::core::u32::MAX)) as u32;
-        Bernoulli { p_int }
+        Bernoulli { p }
     }
 }
 
 impl Distribution<bool> for Bernoulli {
     #[inline]
     fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> bool {
-        // If `p` is constant, this will be evaluated at compile-time.
-        rng.gen::<u32>() <= self.p_int
+        rng.gen::<f64>() <= self.p
     }
 }