* This just uses `gen_bool`, not a different implementation.
* It is implemented for all primitive types. Not sure whether that makes sense. Maybe it is better to only generate `bool` and let the user do the conversion? Having `Distribution<T>` implemented for more than one `T` can make type annotations annoying. If we ever add `impl Distribution<f32> for Normal`, it will probably break a lot of code...