@dhardy 2018-04-19
I don't see how this achieves much, especially the implementations for non-boolean types.

(I know I said I wanted a Bernoulli distribution, but `gen_bool` is that — explicitly implementing `Distribution<T>` doesn't appear to add much.)