@vks 2018-04-19
I think this is more consistent. Similar to how  `Uniform` is redundant to `gen_range`.

If inlining and optimization work, the assert should be only evaluated once, so it could be slightly more efficient in tight loops.