@vks 2018-04-19
It seems the performance difference can be significant:

```
test misc_bernoulli_const            ... bench:       1,165 ns/iter (+/- 7)
test misc_bernoulli_var              ... bench:       2,258 ns/iter (+/- 31)
test misc_gen_bool_const             ... bench:       1,067 ns/iter (+/- 14)
test misc_gen_bool_var               ... bench:       4,370 ns/iter (+/- 72)
```