@vks 2018-04-23
Going to 64 bit using the same method requires special casing 1.0, because of a fun bug in the compiler:

```rust
1.0 * (::core::u64::MAX as f64)) as u64 == u64::MAX
black_box(1.0) * (::core::u64::MAX as f64)) as u64 == 0
```