@pitdicker 2018-04-23
I think you are just hitting the problem that 2^64 - 1 is not representable with `f64`. Can you try multiplying with `2.0.powi(64)` instead?
With the `1.0` probability special-cased to always produce true?