@vks 2018-04-23
How would that help? Note that `::core::u64::MAX as f64` works, but the conversion back to `u64` is Undefined Behavior. `1.0` is probably not the only problematic value, values close to it are also affected.