@pitdicker 2018-04-23
I am pretty sure it does not work that way. `::core::u64::MAX as f64` does not work, as that value is not representable in an `f64` (discussed above). And the only problematic values that may be ub (not read the whole thread, but I just assume all values that are not representable in an `u64`) are `>= 1.0`. If you special-case `1.0` there should be no trouble. 1 ULP below 1.0 maps to a value a couple of thousand less than `u64::MAX`, so values close to it should not be be affected.