@vks 2018-04-23
> ::core::u64::MAX as f64 does not work, as that value is not representable in an f64 (discussed above).

It gives exactly the same result as `2.0.powi(64)` though.