@pitdicker 2018-04-23
> It gives exactly the same result as `2.0.powi(64) - 1.0` though.

This discussion is going a bit difficult as I should just write some code to test things... But `2.0.powi(64) - 1.0` is also not representable. I imagine they both round to either 2^64, or one ULP below that.