@pitdicker 2018-04-24
@vks What do you think of [this]https://github.com/pitdicker/rand/commit/fe69ee55dfa54d853cb21aa06a8d6f80462ff2dd)? The extra branch seems mostly free thanks to branch prediction, and I think this is good and simple enough to put concerns over bias to rest.