@vks 2018-04-24
> @vks What do you think of this? The extra branch seems mostly free thanks to branch prediction, and I think this is good and simple enough to put concerns over bias to rest.

I decided against it, because the extra branch for every case did not seem worth it for getting rid of the tiny bias for a special case. I did not check how big the performance impact would be though.