@pitdicker 2018-04-24
Benchmarks with the branch:
```
test misc_bernoulli_const            ... bench:       2,153 ns/iter (+/- 728)
test misc_bernoulli_var              ... bench:       3,034 ns/iter (+/- 164)
```

Benchmarks without:
```
test misc_bernoulli_const            ... bench:       2,091 ns/iter (+/- 310)
test misc_bernoulli_var              ... bench:       2,964 ns/iter (+/- 307)
```

I mainly suggested all this because @dhardy had concerns about the accuracy of `gen_bool`, and @huonw about bias. As a distribution can move part of the cost to an initialization method, a distribution can be the more accurate variant over `gen_bool` at little cost.