@pitdicker 2018-04-24
> That difference seems to be negligible, I could definitely live with that.

I can already live with the 32-bit accuracy and tiny bias, even more so with this with 64 bit. But as it costs us little, having no easy to point out bias seems like a win to me (no need to defend the choice later...)

@dhardy Interested in making a decisive vote?