@dhardy 2018-04-24
The suggestion I just made eliminates bias, except unfortunately with regards to 1.0.

@pitdicker your suggestion makes the distribution's memory larger. Why not map 1.0 to `MAX` as @vks has but match that value when sampling? `1 - ε/2` will map to significantly less than `u64::MAX` because `u64` has ~11 bits more precision.