@pitdicker 2018-04-24
> Why not map 1.0 to MAX as @vks has but match that value when sampling?

That is a good idea. Just use a value known to be impossible instead of `Option` to reduce overhead.