@vks 2018-04-24
I added the special case for `p = 1` sampling.