@pitdicker 2018-04-25
Then there was one point of discussion left from the issue: should `gen_bool` produce the same results as the `Bernoulli` distribution? Having something available with more precision, even though usually unnecessary, was my reason for re-opening the issue. While you opened an issue hoping to have the same implementation for both.

The idea is that `gen_bool` with 32 bits of accuracy is good enough for almost all uses, and can be about twice as fast as one using 64 bit.

Because benchmarks with Xorshift can be a bit messy (I think it has some problems with using too many registers or something), I benchmarked here with `StdRng`. This PR:
```
test misc_gen_bool_const             ... bench:       4,258 ns/iter (+/- 10)
test misc_gen_bool_var               ... bench:       5,883 ns/iter (+/- 10)

test misc_bernoulli_const            ... bench:       4,238 ns/iter (+/- 7)
test misc_bernoulli_var              ... bench:       4,301 ns/iter (+/- 8)
```

Before:
```
test misc_gen_bool_const             ... bench:       2,533 ns/iter (+/- 6)
test misc_gen_bool_var               ... bench:       3,234 ns/iter (+/- 14)
```