@vks 2018-04-25
> Then there was one point of discussion left from the issue: should gen_bool produce the same results as the Bernoulli distribution?

I think it should. It is a convenience method (that I would actually prefer to remove), having a different precision would be surprising. If we want a lower precision method, I think it should be implemented as an additional distribution. Then we could decide which one `gen_bool` should use.