@dhardy 2018-04-26
I ran the benchmarks. Is the slowdown to `misc_gen_bool` purely due to increased precision or would a few `#[inline]` / `#[inline(always)]` annotations help?
```
$ export RUSTFLAGS="-C codegen-units=1"
$ cargo +nightly bench --bench misc

# before:
test gen_1k_sample_iter              ... bench:         384 ns/iter (+/- 15) = 2666 MB/s
test misc_gen_bool                   ... bench:       1,397 ns/iter (+/- 45)
test misc_gen_bool_var               ... bench:       4,508 ns/iter (+/- 128)

# now:
test gen_1k_sample_iter              ... bench:         592 ns/iter (+/- 69) = 1729 MB/s
test misc_bernoulli_const            ... bench:       3,168 ns/iter (+/- 156)
test misc_bernoulli_var              ... bench:       3,019 ns/iter (+/- 121)
test misc_gen_bool_const             ... bench:       3,050 ns/iter (+/- 73)
test misc_gen_bool_var               ... bench:       4,233 ns/iter (+/- 202)
```

I don't get why `gen_1k_sample_iter` changes; it shouldn't be affected.

@pitdicker would you like to review before this gets merged? I think it's nearly ready.