@pitdicker 2018-04-27
>I ran the benchmarks. Is the slowdown to misc_gen_bool purely due to increased precision or would a few #[inline] / #[inline(always)] annotations help?

The slowdown is due to the increased precision.